Criando Controller para novas Funcionalidades

+++ ImplicareApplication/web/GerenciarVaga.jsp
@@ -74,10 +74,16 @@
                                     <tbody>
                                 </table>
                                 <div class="little-container col-lg-3 col-md-12 col-sm-12">
-                                    <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="ResponderCandidatos" href="">Responder Candidatos</a>
-                                    <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="GerenciarCandidatos" href="">Gerenciar Candidatos</a>
-                                    <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="EditarVaga" href="ImplicareServlet?acao=PesquisarVaga&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Editar Vaga</a>
-                                    <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="ExcluirVaga" href="ImplicareServlet?acao=ExcluirVaga&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Excluir Vaga</a>
+                                    <%if(ListaVaga.get(i).getStatus_Vaga() == 'A'){%>
+                                        <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="ResponderCandidatos" href="ImplicareServlet?acao=ListarCandidatoVaga&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Responder Candidatos</a>
+                                        <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="GerenciarCandidatos" href="ImplicareServlet?acao=ListarCandidatoVagaAceito&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Gerenciar Candidatos</a>
+                                        <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="EditarVaga" href="ImplicareServlet?acao=PesquisarVaga&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Editar Vaga</a>
+                                        <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="ExcluirVaga" href="ImplicareServlet?acao=ExcluirVaga&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Excluir Vaga</a>
+                                    <%} else {%>
+                                            <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="GerenciarCandidatos" href="ImplicareServlet?acao=ListarCandidatoVagaAceito&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Gerenciar Candidatos</a>
+                                            <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="EditarVaga" href="ImplicareServlet?acao=PesquisarVaga&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Editar Vaga</a>
+                                            <a type="submit" class="btn-gerenciar-vaga btn btn-primary1" id="ExcluirVaga" href="ImplicareServlet?acao=ExcluirVaga&Seq_Vaga=<%=ListaVaga.get(i).getSeq_Vaga()%>">Excluir Vaga</a>
+                                        <%}%>
                                 </div>
                             </div>
                         </div>
